Updated PebbleManager to new data sending schema
Updated PebbleManager class to accommodate new run sending schema. Now, the pebble will first send an open message, the phone will return a UUID for the run, and then the pebble will proceed to send the run information. Specifically, it will send the duration and number of laps, followed by the lap times, ending with a close message. This will make sure that the run will only be saved under good communications. The app also now has a more complex method of sending the initial data. If it receives a request from the pebble for the initial data (in the form of key REQUEST_INITIAL), then it will send said data.