Final corrections for android MVP
Made final corrections for the android side of the MVP. All these corrections are in the PebbleManager. Corrected an error where the UUID_DEFINE PebbleDictionary was created but never sent, and corrected the RUN_CLOSE action, as it was performing an invalid cast from Time to long. Also added acknowledge message so that it is sent back to the Pebble