Added Google Analytics and button highlight
Added class AnalyticsManager to manage all Google Analytics code. Simply an initial test. Propagated code through the activities. Also corrected the buttons so that they have a hilight when pressed. They are rounded on the edges that are not pressed against the activity edge (top for the bottom button and bottom for the top button). Also miscelaneous minor changes (PebbleManager constructor is now private, RunManager now only needs RunActivity and PreciseChronometer as arguments, etc.). THIS IS CONSIDERED VERSION CODE 2 (1.1)